Don't abuse PRNG -> fix identifier generation when building on Windows
Pseudorandom number generators need to be initialized only once.

Using the system time for this is Generally Fine, but re-initializing
the PRNG in a tight loop every time a random number is needed
significantly raises the odds that the same number sequence (or even
the same number) is returned every time.

Chances for this to happen on Windows are apparently much higher than
on Linux. Not sure why this is the case, my best guess is that the
clock used for Go's `time.Now()´ runs with coarser granularity.